+++ source/classes/Adventure.js
@@ -112,10 +112,11 @@ class Adventure {
 	getBaseScore() {
 		const livesScore = this.lives * 10;
 		const goldScore = Math.floor(Math.log10(this.peakGold)) * 5;
+		const artifactMultiplier = 1 + (this.getArtifactCount("Floating Multiplier") / 4);
 		return {
 			livesScore,
 			goldScore,
-			total: this.score + this.depth + livesScore + goldScore
+			total: Math.floor((this.score + this.depth + livesScore + goldScore) * artifactMultiplier)
 		};
 	}